Fixed most of the issues that were annoying me for quite some time
1. Put app-wide constants in a single file app_admin, that also
contains a traceback function to warn admin that something is
wrong.
2. Split robust_eSearch_query from eUtils.py to create a function
get_hit_count that allows to perform cheap PubMed hit count.
3. Limited the number of hits per mail to 40. Wrote a message in
the PubCron mail to tell the user when that happens.
4. Wrote another warning in the PubCron mail telling the user
that their query is too big (>200 hits) and that hits have been
ignored.
5. Got rid of one of the warning windows when sending feedback by
sending the form to https://... This alone did the trick, I have
no clue whether the connection is really encrypted or not.
6. Replaced the term update by hit count, which is faster.
7. Added an error traceback mail to admin when term update fails
for unknown reason.